refactor nutrition_facts, add ShoppingList, add tags
Every QuantifiedFood can also hold a list of tags. These are used to
guide different aggregation procedures on QuantifiedFoods.

Previously the only aggregate in Nutcalc was NutritionFacts, but this
patch introduces the ShoppingList aggregate, which is guided by tags.
When defining a CompoundFood, we give a list of quantified foods. Each
of these QuantifiedFoods now holds a list of tags. The tags "buy" and
"use" are meaningful for the `shopping_list` aggregation. Items marked
'buy' are added to the shopping list. Unmarked items are traversed
recursively to find other 'buy' items. Items marked 'use' block
traversal.

In a future patch it would be very very interesting to see if
user-defined aggregates and traversals can be implemented, generalizing
NutritionFacts and ShoppingList to something in 'userspace'. Then the
recursive functions `shopping_list` and `nutrition_facts` could be
implemented as libraries rather than built-in to the system.

This patch also refactors `nutrition_facts`. Previously, this function's
cases were "spread out" among the Nutrient and CompoundFood classes.
Instead, I hoisted the cases into a single recursive function and moved
it into the interpret.py module. Now `nutrition_facts` and
`shopping_list` function the same way: take a QuantifiedFood as input
and proceed by cases, analyzing the tags and the quantity to do the
calculation, returning an appropriate aggregate.